pyproject.toml
[project]
name = "quietlelouch-content-service"
version = "0.1.0"
description = ""
authors = [
    {name = "wokie",email = "contextyoo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "faststream (>=0.5.40,<0.6.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "dishka (>=1.5.3,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "aioboto3 (>=14.3.0,<15.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
